requirements.txt
faster-whisper==1.0.3
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2
huggingface_hub==0.19.4
numpy==1.26.4